bug: still trying to troubleshoot

+++ 0x02-python_async_comprehension/0-async_generator.py
@@ -12,7 +12,7 @@ import random
 from typing import Generator
 
 
-async def async_generator() -> Generator[float, None, None]:
+async def async_generator() -> Generator[int, None, None]:
     """
     This coroutine yields 10 random float numbers between 0 and 10.
     """
